requirements.txt
streamlit
pandas
requests
lxml
pytz
python-dateutil
entsoe-py
altair
supabase
